+++ src/Pim/Bundle/FlexibleEntityBundle/Form/Type/EmailCollectionType.php
@@ -25,7 +25,6 @@ class EmailCollectionType extends CollectionAbstract
             'collection',
             array(
                 'type'           => new EmailType(),
-
                 'allow_add'      => true,
                 'allow_delete'   => true,
                 'by_reference'   => false,