+++ src/Pim/Bundle/ImportExportBundle/Reader/File/YamlReader.php
@@ -4,7 +4,6 @@ namespace Pim\Bundle\ImportExportBundle\Reader\File;
 
 use Symfony\Component\Yaml\Yaml;
 use Oro\Bundle\BatchBundle\Item\ItemReaderInterface;
-use Pim\Bundle\ImportExportBundle\Reader\File\FileReader;
 
 /**
  * Yaml reader
@@ -64,13 +63,12 @@ class YamlReader extends FileReader implements ItemReaderInterface
     }
 
     /**
-     * Get the code field
-     *
-     * @return string
+     * Get file path
+     * @return string $filePath
      */
-    public function getCodeField()
+    public function getFilePath()
     {
-        return $this->codeField;
+        return $this->filePath;
     }
 
     /**
@@ -85,6 +83,16 @@ class YamlReader extends FileReader implements ItemReaderInterface
         $this->codeField = $codeField;
 
         return $this;
+    }
+
+    /**
+     * Get the code field
+     *
+     * @return string
+     */
+    public function getCodeField()
+    {
+        return $this->codeField;
     }
 
     /**
@@ -121,19 +129,32 @@ class YamlReader extends FileReader implements ItemReaderInterface
         }
 
         if ($this->homogenize) {
-            $labels = array();
-            foreach ($fileData as $row) {
-                $labels = array_unique(array_merge($labels, array_keys($row)));
-            }
-            foreach ($fileData as $key => $row) {
-                $missing = array_diff($labels, array_keys($row));
-                foreach ($missing as $label) {
-                    $fileData[$key][$label] = null;
-                }
-            }
+            $fileData = $this->homogenizeData($fileData);
         }
 
         return $this->multiple ? array($fileData) : $fileData;
+    }
+
+    /**
+     * Homogenize the read data
+     *
+     * @param  array $data
+     * @return array
+     */
+    protected function homogenizeData($data)
+    {
+        $labels = array();
+        foreach ($data as $row) {
+            $labels = array_unique(array_merge($labels, array_keys($row)));
+        }
+        foreach ($data as $key => $row) {
+            $data[$key] += array_fill_keys(
+                array_diff($labels, array_keys($row)),
+                null
+            );
+        }
+
+        return $data;
     }
 
     /**