+++ src/Pim/Bundle/ImportExportBundle/Tests/Unit/Reader/CachedReaderTest.php
@@ -15,7 +15,7 @@ class CachedReaderTest extends \PHPUnit_Framework_TestCase
 {
     public function testRead()
     {
-        $reader = new CachedReader;
+        $reader = new CachedReader();
         $data = array(
             'row1' => array('key1' => 'value1', 'key2' => 'value2'),
             'row2' => array('key1' => 'value3', 'key2' => 'value4'),
@@ -38,5 +38,11 @@ class CachedReaderTest extends \PHPUnit_Framework_TestCase
         }
 
         $this->assertNull($reader->read());
+    }
+
+    public function testGetConfigurationFields()
+    {
+        $reader = new CachedReader;
+        $this->assertEquals(array(), $reader->getConfigurationFields());
     }
 }