+++ src/Pim/Bundle/ImportExportBundle/Tests/Unit/Transformer/Guesser/EntityGuesserTest.php
@@ -92,7 +92,7 @@ class EntityGuesserTest extends GuesserTestCase
                 $this->returnValue(array('targetEntity' => 'target_entity', 'type' => ClassMetadataInfo::MANY_TO_MANY))
             );
 
-        $repository = new \stdClass;
+        $repository = new \stdClass();
         $this->doctrine->expects($this->once())
             ->method('getRepository')
             ->with($this->equalTo('target_entity'))